refactor(oauth): move redirect helpers to module
