+++ benefits/oauth/views.py
@@ -2,10 +2,9 @@ import logging
 
 from django.shortcuts import redirect
 from django.urls import reverse
-from django.utils.http import urlencode
 
 from benefits.core import session
-from . import analytics, client
+from . import analytics, client, redirects
 
 
 logger = logging.getLogger(__name__)
@@ -24,7 +23,7 @@ def login(request):
     analytics.started_sign_in(request)
 
     route = reverse(ROUTE_AUTH)
-    redirect_uri = _generate_redirect_uri(request, route)
+    redirect_uri = redirects.generate_redirect_uri(request, route)
 
     logger.debug(f"OAuth authorize_redirect with redirect_uri: {redirect_uri}")
 
@@ -62,13 +61,13 @@ def logout(request):
     session.logout(request)
 
     route = reverse(ROUTE_POST_LOGOUT)
-    redirect_uri = _generate_redirect_uri(request, route)
+    redirect_uri = redirects.generate_redirect_uri(request, route)
 
     logger.debug(f"OAuth end_session_endpoint with redirect_uri: {redirect_uri}")
 
     # send the user through the end_session_endpoint, redirecting back to
     # the post_logout route
-    return _deauthorize_redirect(token, redirect_uri)
+    return redirects.deauthorize_redirect(token, redirect_uri)
 
 
 def post_logout(request):
@@ -78,34 +77,3 @@ def post_logout(request):
 
     origin = session.origin(request)
     return redirect(origin)
-
-
-def _deauthorize_redirect(token, redirect_uri):
-    """Helper implements OIDC signout via the `end_session_endpoint`."""
-
-    # Authlib has not yet implemented `end_session_endpoint` as the OIDC Session Management 1.0 spec is still in draft
-    # See https://github.com/lepture/authlib/issues/331#issuecomment-827295954 for more
-    #
-    # The implementation here was adapted from the same ticket: https://github.com/lepture/authlib/issues/331#issue-838728145
-    oauth_client = client.instance()
-
-    metadata = oauth_client.load_server_metadata()
-    end_session_endpoint = metadata.get("end_session_endpoint")
-
-    params = dict(id_token_hint=token, post_logout_redirect_uri=redirect_uri)
-    encoded_params = urlencode(params)
-    end_session_url = f"{end_session_endpoint}?{encoded_params}"
-
-    return redirect(end_session_url)
-
-
-def _generate_redirect_uri(request, redirect_path):
-    redirect_uri = str(request.build_absolute_uri(redirect_path)).lower()
-
-    # this is a temporary hack to ensure redirect URIs are HTTPS when the app is deployed
-    # see https://github.com/cal-itp/benefits/issues/442 for more context
-    # this follow-up is needed while we address the hosting architecture
-    if not redirect_uri.startswith("http://localhost"):
-        redirect_uri = redirect_uri.replace("http://", "https://")
-
-    return redirect_uri
